+++ cli/src/cli/mod.rs
@@ -3,6 +3,7 @@ pub mod client;
 pub mod deposit;
 pub mod error;
 pub mod get;
+pub mod history;
 pub mod send;
 pub mod sync;
 pub mod utils;
