+++ cli/src/main.rs
@@ -6,7 +6,8 @@ use intmax2_cli::{
         claim::claim_withdrawals,
         deposit::deposit,
         error::CliError,
-        get::{balance, claim_status, history, mining_list, withdrawal_status},
+        get::{balance, claim_status, mining_list, withdrawal_status},
+        history::history,
         send::send_transfers,
         sync::{sync_claims, sync_withdrawals},
         withdrawal::send_withdrawal,
